classify/requirements.txt
openai
pandas
scikit-learn
numpy
python-dotenv